refactor: use single `renderBuiltUrl` function
