+++ packages/vite/package.json
@@ -57,7 +57,6 @@
     "std-env": "^3.9.0",
     "ufo": "^1.6.1",
     "unenv": "^2.0.0-rc.17",
-    "unplugin": "^2.3.5",
     "vite": "^6.3.5",
     "vite-node": "^3.2.4",
     "vite-plugin-checker": "^0.9.3",
